lesson 12

+++ app/controllers/decks_controller.rb
@@ -2,7 +2,7 @@ class DecksController < ApplicationController
   before_action :load_deck, only: [:edit, :update, :destroy, :current]
 
   def index
-    @decks = current_user.decks.ordered
+    @decks = logged_in? ? current_user.decks.ordered : []
   end
 
   def new
@@ -10,12 +10,17 @@ class DecksController < ApplicationController
   end
 
   def create
+    if logged_in?
     @deck = Deck.new deck_params
-    @deck.user = current_user
-    if @deck.save
-      redirect_to decks_path, notice: t(:deck_created)
+      @deck.user = current_user
+      if @deck.save
+        redirect_to decks_path, notice: t(:deck_created)
+      else
+        render :new
+      end
     else
-      render :new
+      redirect_to decks_path
+      flash[:error] = 'Залогиньтесь!'
     end
   end
 
@@ -23,27 +28,51 @@ class DecksController < ApplicationController
   end
 
   def update
-    if @deck.update deck_params
-      redirect_to edit_deck_path @deck
+    if logged_in?
+      if @deck.user_id == current_user.id
+        if @deck.update deck_params
+          redirect_to edit_deck_path @deck
+        else
+          render :edit
+        end
+      else
+        redirect_to edit_deck_path @deck
+        flash[:error] = 'Вы не можете редактировать чужую колоду!'
+      end
     else
-      render :edit
+      redirect_to edit_deck_path @deck
+      flash[:error] = 'Залогиньтесь!'
     end
   end
 
   def destroy
-    if @deck.destroy
-      redirect_to decks_path, notice: t(:deck_deleted)
+    if logged_in?
+      if @deck.user_id == current_user.id
+        if @deck.destroy
+          redirect_to decks_path, notice: t(:deck_deleted)
+        else
+          redirect_to decks_path
+          flash[:error] = t(:deck_not_deleted)
+        end
+      else
+        redirect_to decks_path
+        flash[:error] = 'Вы не можете удалить чужую колоду!'
+      end
     else
       redirect_to decks_path
-      flash[:error] = t(:deck_not_deleted)
+      flash[:error] = 'Залогиньтесь!'
     end
   end
 
   def current
-    if @deck.set_current current_user
-      render json: { status: 'ok' }
+    if logged_in? && @deck.user_id == current_user.id
+      if @deck.set_current current_user
+        render json: { status: 'ok' }
+      else
+        render json: { status: 'fail' }
+      end
     else
-      render json: { status: 'fail' }
+        render json: { status: 'fail' }
     end
   end
